Figured out how webpack works!  DEATH TO THE MEGA FILE!
