+++ src/main/js/cardGrid.js
@@ -1,0 +1,39 @@
+const React = require('react');
+
+import { withStyles } from "@material-ui/core/styles/index";
+import Grid from '@material-ui/core/Grid';
+import { NewRoundCard } from './newRoundCard';
+import { NewFixedStatCombatantCard } from "./newFixedStatCombatantCard";
+import { NewNpcsFromTemplateCard } from "./newNpcsFromTemplateCard";
+
+const gridStyles = theme => ({
+  root: {
+    flexGrow: 1,
+    marginTop: theme.spacing.unit * 4,
+  },
+  paper: {
+    padding: theme.spacing.unit * 2,
+    textAlign: 'center',
+    color: theme.palette.text.secondary,
+  },
+});
+class CardGridUnstyled extends React.Component {
+  render() {
+    return (
+      <div className={this.props.classes.root}>
+        <Grid container spacing={24}>
+          <Grid item xs={12} sm={4}>
+            <NewRoundCard/>
+          </Grid>
+          <Grid item xs={12} sm={4}>
+            <NewFixedStatCombatantCard/>
+          </Grid>
+          <Grid item xs={12} sm={4}>
+            <NewNpcsFromTemplateCard/>
+          </Grid>
+        </Grid>
+      </div>
+    )
+  }
+}
+export const CardGrid = withStyles(gridStyles)(CardGridUnstyled);
