+++ src/main/js/newFixedStatCombatantCard.js
@@ -1,0 +1,53 @@
+const React = require('react');
+
+import { withStyles } from "@material-ui/core/styles/index";
+import Card from '@material-ui/core/Card';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import Button from '@material-ui/core/Button';
+import Typography from '@material-ui/core/Typography';
+
+
+const cardStyle = theme => ({
+  root: {
+    padding: theme.spacing.unit * 2,
+  },
+  card: {
+    minWidth: 275,
+  },
+  bullet: {
+    display: 'inline-block',
+    margin: '0 2px',
+    transform: 'scale(0.8)',
+  },
+  title: {
+    marginBottom: 16,
+    fontSize: 14,
+  },
+  pos: {
+    marginBottom: 12,
+  },
+});
+
+class NewFixedStatCombatantCardUnstyled extends React.Component {
+  render() {
+    return (
+      <div className={this.props.classes.root}>
+        <Card className={this.props.classes.card}>
+          <CardContent>
+            <Typography className={this.props.classes.title} color="textSecondary">
+              New Fixed Stat Combatant Wizard
+            </Typography>
+            <Typography className={this.props.classes.pos} color="textSecondary">
+              This wizard will help you create either a new character or npc and add it to combat
+            </Typography>
+          </CardContent>
+          <CardActions>
+            <Button size="small">New Round</Button>
+          </CardActions>
+        </Card>
+      </div>
+    )
+  }
+}
+export const NewFixedStatCombatantCard = withStyles(cardStyle)(NewFixedStatCombatantCardUnstyled);
